Landing page now pulls from firebase listings. To be merged with Add listing

+++ app/src/main/java/com/mobdeve/s15/animall/ListingModel.kt
@@ -6,7 +6,7 @@ class ListingModel(
     val name: String, //TODO: Update to link to user in db
     val preferredLocation: String,
     val seller: String,
-    val stock: Int,
+    val stock: Long,
     val unitPrice: Double,
     val photos: Array<String> //TODO: Update to links in the future
 )
